requirements.txt
dvc
opencv-python
opencv-contrib-python
parameterized
Pillow
streamlit
ultralytics